+++ src/pages/postManager/ui/PostManager.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 
-import { Plus, Search } from 'lucide-react';
+import { Plus } from 'lucide-react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
 import {
@@ -9,7 +9,6 @@ import {
   useLimit,
   useLoading,
   usePostActions,
-  usePostFilterActions,
   usePosts,
   useSelectedPost,
   useSelectedTag,
@@ -21,12 +20,10 @@ import {
   useShowPostDetailDialog,
   useNewComment,
   useUserActions,
-  useSearchQuery,
 } from '@/shared/model/store';
 
 import { Card, CardContent, CardHeader, CardTitle } from '@/shared/ui';
 import { Button } from '@/shared/ui';
-import { Input } from '@/shared/ui';
 
 import { PostDetail } from './PostDetail';
 import { PostTable } from './PostTable';
@@ -41,6 +38,7 @@ import { SortByOrder } from '@/feature/sortByOrder';
 import { Pagination } from '@/feature/pagination';
 import { UserInfo } from '@/feature/userInfo';
 import { useQueryParams } from '@/shared/hooks/useQueryParams';
+import { SearchPost } from '@/feature/searchPost';
 
 export const PostsManager = () => {
   useQueryParams();
@@ -49,7 +47,6 @@ export const PostsManager = () => {
   const location = useLocation();
   const queryParams = new URLSearchParams(location.search);
 
-  // 전역상태
   const posts = usePosts();
   const selectedPost = useSelectedPost();
   const loading = useLoading();
@@ -60,9 +57,6 @@ export const PostsManager = () => {
   const limit = useLimit();
   const sortBy = useSortBy();
   const sortOrder = useSortOrder();
-  const searchQuery = useSearchQuery();
-
-  const { setSearchQuery } = usePostFilterActions();
 
   const selectedTag = useSelectedTag();
 
@@ -125,24 +119,6 @@ export const PostsManager = () => {
     } catch (error) {
       console.error('태그 가져오기 오류:', error);
     }
-  };
-
-  // 게시물 검색
-  const searchPosts = async () => {
-    if (!searchQuery) {
-      fetchPosts();
-      return;
-    }
-    setLoading(true);
-    try {
-      const response = await fetch(`/api/posts/search?q=${searchQuery}`);
-      const data = await response.json();
-      setPosts(data.posts);
-      setTotal(data.total);
-    } catch (error) {
-      console.error('게시물 검색 오류:', error);
-    }
-    setLoading(false);
   };
 
   // 태그별 게시물 가져오기
@@ -303,19 +279,7 @@ export const PostsManager = () => {
           <div className='flex flex-col gap-4'>
             {/* 검색 및 필터 컨트롤 */}
             <div className='flex gap-4'>
-              <div className='flex-1'>
-                <div className='relative'>
-                  <Search className='absolute left-2 top-2.5 h-4 w-4 text-muted-foreground' />
-                  <Input
-                    placeholder='게시물 검색...'
-                    className='pl-8'
-                    value={searchQuery}
-                    onChange={(e) => setSearchQuery(e.target.value)}
-                    onKeyDown={(e) => e.key === 'Enter' && searchPosts()}
-                  />
-                </div>
-              </div>
-
+              <SearchPost />
               <SortByTag />
               <SortByValue />
               <SortByOrder />
